+++ src/io/mod.rs
@@ -7,3 +7,4 @@ pub(crate) mod geozero;
 pub(crate) mod native;
 #[cfg(feature = "parquet")]
 pub mod parquet;
+pub mod wkb;
