+++ src/io/native/mod.rs
@@ -1,3 +1,2 @@
 #[cfg(feature = "geos")]
 pub mod geos;
-pub mod wkb;
