+++ src/io/wkb/reader/maybe_multi_point.rs
@@ -1,9 +1,9 @@
 use crate::algorithm::native::eq::multi_point_eq;
 use crate::geo_traits::MultiPointTrait;
-use crate::io::native::wkb::multipoint::WKBMultiPoint;
-use crate::io::native::wkb::point::WKBPoint;
-// use crate::io::native::wkb::Point::WKBPoint;
-// use crate::io::native::wkb::multiPoint::WKBMultiPoint;
+use crate::io::wkb::reader::multipoint::WKBMultiPoint;
+use crate::io::wkb::reader::point::WKBPoint;
+// use crate::io::wkb::Point::WKBPoint;
+// use crate::io::wkb::multiPoint::WKBMultiPoint;
 use std::iter::Cloned;
 use std::slice::Iter;
 
@@ -80,7 +80,7 @@ impl<'a> MultiPointTrait<'a> for &WKBMaybeMultiPoint<'a> {
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::io::native::wkb::geometry::Endianness;
+    use crate::io::wkb::reader::geometry::Endianness;
     use crate::test::multipoint::mp0;
     use crate::test::point::p0;
     use geozero::{CoordDimensions, ToWkb};
