+++ src/io/wkb/reader/maybe_multipolygon.rs
@@ -1,7 +1,7 @@
 use crate::algorithm::native::eq::multi_polygon_eq;
 use crate::geo_traits::MultiPolygonTrait;
-use crate::io::native::wkb::multipolygon::WKBMultiPolygon;
-use crate::io::native::wkb::polygon::WKBPolygon;
+use crate::io::wkb::reader::multipolygon::WKBMultiPolygon;
+use crate::io::wkb::reader::polygon::WKBPolygon;
 use std::iter::Cloned;
 use std::slice::Iter;
 
@@ -78,7 +78,7 @@ impl<'a> MultiPolygonTrait<'a> for &WKBMaybeMultiPolygon<'a> {
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::io::native::wkb::geometry::Endianness;
+    use crate::io::wkb::reader::geometry::Endianness;
     use crate::test::multipolygon::mp0;
     use crate::test::polygon::p0;
     use geozero::{CoordDimensions, ToWkb};
